moved add account modal logic to its component
