+++ src/app/home/components/account-modal/account-modal.component.html
@@ -1,3 +1,86 @@
-<p>
-  account-modal works!
-</p>
+<ion-header>
+  <ion-toolbar>
+    <ion-title>{{ "ADD_ACCOUNT_MODAL.TITLE" | translate }}</ion-title>
+    <ion-buttons slot="end">
+      <ion-button (click)="dismiss()"><ion-icon slot="icon-only" name="close-outline"></ion-icon></ion-button>
+    </ion-buttons>
+  </ion-toolbar>
+</ion-header>
+<ion-content class="ion-padding">
+  <ng-container *ngIf="isScanActive;else manualForm">
+    <ngx-scanner-qrcode #qrscanner="scanner" (event)="onQRCodeScanned($event, qrscanner)"
+      [config]="qrScannerOpts"></ngx-scanner-qrcode>
+    <ion-grid>
+      <ion-row>
+        <ion-col size="auto">
+        </ion-col>
+        <ion-col></ion-col>
+        <ion-col size="auto">
+          <ion-button size="large" (click)="cycleCamera()">
+            <ion-icon slot="icon-only" name="sync-outline"></ion-icon>
+          </ion-button>
+        </ion-col>
+      </ion-row>
+      <ion-row>
+        <ion-col size="12">
+          <ion-button size="large" expand="block" (click)="manualInputAction()">
+            <ion-icon slot="start" name="create-outline"></ion-icon>
+            {{ "ADD_ACCOUNT_MODAL.MANUAL_INPUT" | translate }}
+          </ion-button>
+        </ion-col>
+      </ion-row>
+    </ion-grid>
+  </ng-container>
+  <ng-template #manualForm>
+    <ion-button size="large" expand="block" (click)="scanCode()">
+      <ion-icon slot="start" name="qr-code-outline"></ion-icon>
+      {{ "ADD_ACCOUNT_MODAL.SCAN_QR_CODE" | translate }}
+    </ion-button>
+    <form [formGroup]="validations_form" (ngSubmit)="createAccount(validations_form.value)">
+      <ion-item>
+        <ion-input label="{{ 'ADD_ACCOUNT_MODAL.LABEL' | translate }}" labelPlacement="stacked" formControlName="label"
+          placeholder="{{ 'ADD_ACCOUNT_MODAL.EXAMPLI_GRATIA_SHORT' | translate }} Google:teste@gmail.com"></ion-input>
+      </ion-item>
+      <ion-item>
+        <ion-input label="{{ 'ADD_ACCOUNT_MODAL.SECRET_KEY' | translate }}" labelPlacement="stacked"
+          formControlName="secret"
+          placeholder="{{ 'ADD_ACCOUNT_MODAL.EXAMPLI_GRATIA_SHORT' | translate }} QAPERTPEO123"></ion-input>
+      </ion-item>
+      <ion-item>
+        <ion-input label="{{ 'ADD_ACCOUNT_MODAL.TOKEN_SIZE' | translate }}" labelPlacement="stacked"
+          formControlName="tokenLength"
+          placeholder="{{ 'ADD_ACCOUNT_MODAL.EXAMPLI_GRATIA_SHORT' | translate }} 6"></ion-input>
+      </ion-item>
+      <ion-item>
+        <ion-input label="{{ 'ADD_ACCOUNT_MODAL.TOKEN_INTERVAL' | translate }}" labelPlacement="stacked"
+          formControlName="interval"
+          placeholder="{{ 'ADD_ACCOUNT_MODAL.EXAMPLI_GRATIA_SHORT' | translate }} 30"></ion-input>
+      </ion-item>
+      <ion-item>
+        <ion-label position="stacked">{{ "ADD_ACCOUNT_MODAL.LOGO" | translate }}</ion-label>
+        <ion-searchbar placeholder="{{ 'SEARCH' | translate }}" class="ion-padding-top" [debounce]="250"
+          [value]="draftLogoSearchTxt" (ionInput)="handleSearchLogo($event)"></ion-searchbar>
+        <ion-grid class="full-row">
+          <ion-row class="full-row">
+            <ion-col size-xs="6" size-sm="4" size="4">
+              <ion-card class="logo-card" (click)="selectLogo('')" [ngClass]="{ 'selected-logo': !draftLogoURL }">
+                <ion-img class="ion-padding logo-img" [src]="'assets/icon/favicon.png'"></ion-img>
+                <p>{{ "ADD_ACCOUNT_MODAL.NO_LOGO" | translate }}</p>
+              </ion-card>
+            </ion-col>
+            <ion-col size-xs="6" size-sm="4" *ngFor="let logoResult of searchLogoResults">
+              <ion-card class="logo-card" (click)="selectLogo(logoResult)"
+                [ngClass]="{ 'selected-logo': logoResult === draftLogoURL }">
+                <ion-img class="ion-padding ion-margin logo-img" [src]="logoResult"></ion-img>
+              </ion-card>
+            </ion-col>
+          </ion-row>
+        </ion-grid>
+        <p *ngIf="searchLogoResults.length > 0">Logos provided by <a href="https://brandfetch.com">Brandfetch</a>
+        </p>
+      </ion-item>
+      <ion-button expand="block" [disabled]="!validations_form.valid" type="submit">{{ "ADD_ACCOUNT_MODAL.SAVE" |
+        translate }}</ion-button>
+    </form>
+  </ng-template>
+</ion-content>
