+++ src/app/home/components/account-modal/account-modal.component.scss
@@ -1,0 +1,20 @@
+.full-row {
+  width: 100%;
+}
+
+.selected-logo {
+  border: 3px solid var(--ion-color-danger);
+}
+
+.logo-img {
+  background-color: whitesmoke;
+  box-shadow: 0 0 3px 0 rgba(0, 0, 0, 0.2);
+}
+
+.logo-card {
+  aspect-ratio: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
