+++ src/app/home/home.page.scss
@@ -40,26 +40,6 @@ ion-content {
   height: 50%;
 }
 
-.logo-card {
-  aspect-ratio: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.logo-img {
-  box-shadow: 0 0 3px 0 rgba(0, 0, 0, 0.2);
-}
-
-.selected-logo {
-  border: 3px solid var(--ion-color-danger);
-}
-
-.full-row {
-  width: 100%;
-}
-
 app-account-list.landscape {
   background-color: var(--ion-item-background, var(--ion-background-color, #fff));
 }
