+++ src/assets/i18n/en.json
@@ -49,7 +49,8 @@
       "ERROR_CAMERA_HEADER": "Camera unavailable",
       "ERROR_CAMERA_MESSAGE": "<p>Camera access is required to scan QR codes, but it seems your device either lacks a camera or camera permissions were denied. </p><p>Please check your device/permissions settings and try again.</p>",
       "INVALID_QR_CODE": "Invalid QR Code",
-      "INVALID_SESSION": "Invalid session"
+      "INVALID_SESSION": "Invalid session",
+      "INVALID_FIELDS": "Invalid account data"
     },
     "EXAMPLI_GRATIA_SHORT": "e.g.",
     "LABEL": "Label",
